requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-limiter==3.5.0
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
google-api-python-client==2.108.0
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
httpx==0.24.1
groq==0.11.0
requests==2.31.0
gunicorn==21.2.0